+++ tests/test_cpp.py
@@ -2,21 +2,21 @@
 
 # Copyright (c) 2019, IRIS-HEP
 # All rights reserved.
-# 
+#
 # Redistribution and use in source and binary forms, with or without
 # modification, are permitted provided that the following conditions are met:
-# 
+#
 # * Redistributions of source code must retain the above copyright notice, this
 #   list of conditions and the following disclaimer.
-# 
+#
 # * Redistributions in binary form must reproduce the above copyright notice,
 #   this list of conditions and the following disclaimer in the documentation
 #   and/or other materials provided with the distribution.
-# 
+#
 # * Neither the name of the copyright holder nor the names of its
 #   contributors may be used to endorse or promote products derived from
 #   this software without specific prior written permission.
-# 
+#
 # THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
 # AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
 # IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
@@ -85,12 +85,12 @@ class Test(unittest.TestCase):
 
     def test_cpp_init(self):
         # This depends on all setters working,
-        # as well as str, repr, and getitem methods
+        # as well as str and getitem methods
         a = numpy.array([1, 4, 1, 0])
         b = numpy.array([2, 9, 5, 10])
         c = numpy.arange(10, dtype=numpy.int16)
         test = awkward.cpp.JaggedArray(a, b, c)
-        assert repr(test) == "<JaggedArray [[1] [4 5 6 7 8] [1 2 3 4] [0 1 2 3 4 5 6 7 8 9]]>"
+        assert str(test) == "[[1] [4 5 6 7 8] [1 2 3 4] [0 1 2 3 4 5 6 7 8 9]]"
 
     def test_cpp_getitem_neg01(self):
         a = numpy.array([1, 2, 3, 4])
@@ -119,4 +119,3 @@ class Test(unittest.TestCase):
                 raise
             thrown = True
         assert thrown
-
